html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-source-into-empty-video.html

<!doctype html>
<title>pointer updates (moveBefore source into an empty media element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="holder"><source id="source-x"></div>
<script>
var events = [];
var b = 0;
var x = 0;
</script>
<video id="v"><source id="source-b" onerror="b++; events.push('b')" src="resources/delayed-broken-video.py"></video>
<script>
async_test(function(t) {
  var video = document.getElementById('v');
  var holder = document.getElementById('holder');
  assert_implements(typeof video.moveBefore == 'function', 'ParentNode.moveBefore() is required for this test');

  var sourceX = document.getElementById('source-x');
  sourceX.onerror = function() { x++; events.push('x'); };

  // Move the current candidate out so the video has no children while b is pending.
  holder.moveBefore(document.getElementById('source-b'), null);

  // Move an existing source from another parent into the empty video.
  video.moveBefore(sourceX, null);

  window.onload = t.step_func(function() {
    assert_equals(b, 1, 'error events on b');
    assert_equals(x, 1, 'error events on x');
    assert_array_equals(events, ['b', 'x'], 'processed source order');
    t.done();
  });
});
</script>
